requirements.txt
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
Brotli==1.0.9
colorama==0.4.6
decorator==5.1.1
dj-database-url==2.1.0
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
executing==1.2.0
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.14.0
jedi==0.19.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
matplotlib-inline==0.1.6
model-bakery==1.14.0
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
pytest-testdox==3.1.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rest-framework-simplejwt==0.0.2
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
wcwidth==0.2.6
whitenoise==6.5.0
